requirements.txt
astroid==2.15.5
dill==0.3.6
exceptiongroup==1.1.1
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
pylint==2.17.4
pytest==7.3.1
python-dotenv==1.0.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
websocket-client==1.5.2
wrapt==1.15.0